Update MCR and MCRData contract
uint64 -> uint

+++ contracts/MCRData.sol
@@ -39,7 +39,7 @@ contract MCRData is Iupgradable {
     struct mcr_Data {
         uint32 mcrPercx100;
         uint32 mcrEtherx100;
-        uint64 vFull; //Pool funds
+        uint vFull; //Pool funds
         uint64 date;
     }
 
@@ -190,7 +190,7 @@ contract MCRData is Iupgradable {
     /// @dev Adds details of (Minimum Capital Requirement)MCR.
     /// @param mcrp Minimum Capital Requirement percentage (MCR% * 100 ,Ex:for 54.56% ,given 5456)
     /// @param vf Pool fund value in Ether used in the last full daily calculation from the Capital model.
-    function pushMCRData(uint32 mcrp, uint32 mcre, uint64 vf, uint64 time) onlyInternal {
+    function pushMCRData(uint32 mcrp, uint32 mcre, uint vf, uint64 time) onlyInternal {
         allMCRData.push(mcr_Data(mcrp, mcre, vf, time));
     }
 
@@ -222,7 +222,7 @@ contract MCRData is Iupgradable {
     /// @dev Gets the details of last added MCR.
     /// @return mcrPercx100 Total Minimum Capital Requirement percentage of that month of year(multiplied by 100).
     /// @return vFull Total Pool fund value in Ether used in the last full daily calculation.
-    function getLastMCR() constant returns(uint32 mcrPercx100, uint32 mcrEtherx100, uint64 vFull, uint64 date) {
+    function getLastMCR() constant returns(uint32 mcrPercx100, uint32 mcrEtherx100, uint vFull, uint64 date) {
         return (
             allMCRData[SafeMaths.sub(allMCRData.length, 1)].mcrPercx100,
             allMCRData[SafeMaths.sub(allMCRData.length, 1)].mcrEtherx100,
@@ -244,7 +244,7 @@ contract MCRData is Iupgradable {
     }
 
     /// @dev Gets Pool fund value in Ether used in the last full daily calculation from the Capital model.
-    function getLastVfull() constant returns(uint64 vf) {
+    function getLastVfull() constant returns(uint vf) {
         vf = allMCRData[SafeMaths.sub(allMCRData.length, 1)].vFull;
     }
